% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge.R
\name{cache_as_list}
\alias{cache_as_list}
\title{Serialise a cache to a plain list}
\usage{
cache_as_list(cache)
}
\arguments{
\item{cache}{An environment from \code{\link[=new_cache]{new_cache()}}.}
}
\value{
A named list.
}
\description{
Serialise a cache to a plain list
}
