% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge.R
\name{default_judge_model}
\alias{default_judge_model}
\title{Default model for the judge and the normaliser}
\usage{
default_judge_model()
}
\value{
A model identifier string.
}
\description{
Default model for the judge and the normaliser
}
