% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_confusion}
\alias{plot_confusion}
\title{A confusion matrix as a tile plot}
\usage{
plot_confusion(cc)
}
\arguments{
\item{cc}{The result of \code{\link[=column_confusion]{column_confusion()}}.}
}
\value{
A ggplot.
}
\description{
A confusion matrix as a tile plot
}
