% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{zip_bundle}
\alias{zip_bundle}
\title{Zip an export bundle}
\usage{
zip_bundle(dir, zipfile = paste0(dir, ".zip"))
}
\arguments{
\item{dir}{A bundle directory from \code{\link[=export_bundle]{export_bundle()}}.}

\item{zipfile}{Destination \code{.zip}.}
}
\value{
\code{zipfile}, invisibly.
}
\description{
Zip an export bundle
}
