% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_completeness}
\alias{plot_completeness}
\title{Fill rates, AI against gold}
\usage{
plot_completeness(fills)
}
\arguments{
\item{fills}{A tibble from \code{\link[=fill_rates]{fill_rates()}}.}
}
\value{
A ggplot.
}
\description{
Fill rates, AI against gold
}
